Renamed module to `bq_schema`
