+++ data_pipeline/crossref_event_data/bq_schema.py
